requirements.txt
awesome-slugify==1.6.5
regex==2024.7.24
Unidecode==0.4.21
